feat: added manual transmission with working gears.
Created a Transmission state that tracks the current gear.
If gear is greater than 0, set current transmission state
to IN_GEAR, meaning the object is moving forward. If gear
is below 0, set current transmission state to REVERSE,
meaning the object is moving backwards. Otherwise, the
object is in a NEUTRAL state.

I also added checks to make sure the transmission has
limits. Meaning if you pass in a number when creating the
Transmission object, the gear cannot go above the max_gear
value. It is also the case for reverse, where it cannot be
less than -1.